*** empty log message ***
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@814 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/library/net/local/client_server/unix_server.e
@@ -10,48 +10,53 @@ indexing
 deferred class UNIX_SERVER
 
 inherit
+
 	SERVER
 		redefine
-			in, make
+			in
 		end
 
 feature -- Access
 
-	in: UNIX_SOCKET_STREAM
+	in: UNIX_STREAM_SOCKET
 		-- Receive socket.
 
-	make (a : SOCKET_ADDRESS_UNIX; queueing : INTEGER) is
+	make (a : STRING) is
 		do
-			!!in.make;
-			in.set_address (a);
-			in.bind;
-			in.listen (queueing);
-		end;
+			!!in.make_server (a)
+			if queued = 0 then
+				in.listen (5)
+			else
+				in.listen (queued)
+			end
+		end
 
-	close is
+	cleanup is
 		do
-			in.close;
-			in.unlink;
-		end;
-
+			in.close
+			if in.address /= Void then
+				in.unlink
+			end
+		end
 
 	receive is
 		do
-			outflow ?= in.accept;
-			received ?= socket_retrieved (outflow);
-		end;
+			in.accept
+			outflow ?= in.accepted
+			received ?= retrieved (outflow)
+		end
 	
-	clean_up is
+	close is
 		do
 			if outflow /= Void and then not outflow.is_closed then
-				outflow.close;
-			end;
-		end;
+				outflow.close
+			end
+		end
 
 end -- class UNIX_SERVER
 
 --|----------------------------------------------------------------
---| EiffelBase: library of reusable components for ISE Eiffel 3.
+--| EiffelNet: library of reusable components for ISE Eiffel 3.
 --| Copyright (C) 1986, 1990, 1993, 1994, Interactive Software
 --|   Engineering Inc.
 --| All rights reserved. Duplication and distribution prohibited.
@@ -62,3 +67,4 @@ end -- class UNIX_SERVER
 --| Electronic mail <info@eiffel.com>
 --| Customer support e-mail <eiffel@eiffel.com>
 --|----------------------------------------------------------------
+
